Moved CPU to its own file
